pyproject.toml
[project]
name = "paisa-app"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "motor>=3.6",
    "pymongo>=4.10",
    "python-dotenv>=1.0",
    "bcrypt>=4.0",
    "pyjwt>=2.8",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic[email]>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "mongomock-motor>=0.0.21",
    "pytest-cov>=5.0",
    "pytest-watch>=4.2",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
